chore: PlainDate separate module
