+++ packages/client/src/views/OfficeHome/inExternalValidation/InExternalValidationTab.tsx
@@ -26,10 +26,8 @@ import {
 } from '@client/i18n/messages'
 import { formatUrl } from '@client/navigation'
 import { navigationMessages } from '@client/i18n/messages/views/navigation'
-import {
-  formattedDuration,
-  plainDateToLocalDate
-} from '@client/utils/date-formatting'
+import { plainDateToLocalDate } from '@opencrvs/commons/client'
+import { formattedDuration } from '@client/utils/date-formatting'
 import {
   getSortedItems,
   changeSortedColumn
