+++ packages/client/src/views/OfficeHome/myDrafts/MyDrafts.tsx
@@ -24,11 +24,8 @@ import {
   DRAFT_MARRIAGE_FORM_PAGE
 } from '@client/navigation/routes'
 import { IStoreState } from '@client/store'
-import {
-  formattedDuration,
-  isValidPlainDate,
-  plainDateToLocalDate
-} from '@client/utils/date-formatting'
+import { plainDateToLocalDate } from '@opencrvs/commons/client'
+import { formattedDuration, isValidPlainDate } from '@client/utils/date-formatting'
 import { getDeclarationFullName } from '@client/utils/draftUtils'
 import { EventType } from '@client/utils/gateway'
 import {
